created check_amperdand
