+++ src/parsing/check_syntax.c
@@ -6,11 +6,10 @@
 /*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/20 15:54:29 by abasdere          #+#    #+#             */
-/*   Updated: 2024/01/25 16:50:30 by abasdere         ###   ########.fr       */
+/*   Updated: 2024/01/25 16:56:26 by abasdere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
 #include "parsing.h"
 
 /**
@@ -64,60 +63,6 @@ t_code	check_quotes(char *line)
 		return (error_syntax(C_BAD_USE, "\'", 1));
 	else if (nd % 2)
 		return (error_syntax(C_BAD_USE, "\"", 1));
-	return (C_SUCCESS);
-}
-
-/**
- * @brief Check pipe syntax in line
- *
- * @param line line to check
- * @param pos position in line
- * @return t_code C_SUCCESS or an error
- */
-t_code	check_pipe(char *line, size_t pos)
-{
-	size_t	nb;
-	size_t	i;
-
-	nb = 0;
-	i = pos;
-	if (line[pos + 1] == '&')
-		return (error_syntax(C_BAD_USE, line + pos, 2));
-	while (line[pos + nb] == '|' || line[pos + nb] == '&')
-		nb++;
-	if (nb == 3)
-		return (error_syntax(C_BAD_USE, line + pos + 2, 1));
-	if (nb > 3)
-		return (error_syntax(C_BAD_USE, line + pos + 2, 2));
-	if (pos && i--)
-		while (i > 0 && line[i] != '|' && !ft_isalnum(line[i]))
-			i--;
-	if (!i && !ft_isalnum(line[i]))
-		return (error_syntax(C_BAD_USE, line + pos, nb));
-	i = pos + 1;
-	while (line[i] && line[i] != '|' && !ft_isalnum(line[i]))
-		i++;
-	if (!line[i])
-		return (error_syntax(C_BAD_USE, line + pos, nb));
-	return (C_SUCCESS);
-}
-
-/**
- * @brief Check special chars syntax in line
- *
- * @param line line to check
- * @return t_code C_SUCCESS or an error
- */
-t_code	check_spcl_chars(char *line)
-{
-	size_t	i;
-
-	i = -1;
-	while (line[++i])
-	{
-		if (line[i] == '|' && check_pipe(line, i))
-			return (C_BAD_USE);
-	}
 	return (C_SUCCESS);
 }
